pyproject.toml
[project]
name = "rfq-repl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ag2[openai]>=0.9.6",
    "aiofiles>=24.1.0",
    "black>=25.1.0",
    "fastapi>=0.116.1",
    "flake8>=7.3.0",
    "mammoth>=1.9.1",
    "mypy>=1.17.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pdfplumber>=0.11.7",
    "psycopg2-binary>=2.9.10",
    "pyautogen>=0.10.0",
    "pydantic>=2.11.7",
    "pypdf2>=3.0.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "structlog>=25.4.0",
    "tenacity>=9.1.2",
    "unicorn>=2.1.3",
    "uvicorn>=0.35.0",
    "xlrd>=2.0.2",
]
